Refactors the SPFx app actions
