+++ src/extension.ts
@@ -14,6 +14,7 @@ import { CHAT_PARTICIPANT_NAME, ProjectFileContent } from './constants';
 import { EntraAppRegistration } from './services/actions/EntraAppRegistration';
 import { CopilotActions } from './services/actions/CopilotActions';
 import { ChatTools } from './chat/tools/ChatTools';
+import { SpfxAppCLIActions } from './services/actions/SpfxAppCLIActions';
 
 
 export async function activate(context: vscode.ExtensionContext) {
@@ -32,6 +33,7 @@ export async function activate(context: vscode.ExtensionContext) {
 	Dependencies.registerCommands();
 	Scaffolder.registerCommands();
 	CliActions.registerCommands();
+	SpfxAppCLIActions.registerCommands();
 	EntraAppRegistration.registerCommands(context);
 	CopilotActions.registerCommands();
